Fix: Fix numeric conversion warnings

+++ src/lib/pca9685.cpp
@@ -42,7 +42,7 @@ std::uint8_t PCA9685::num_servos() {
 // Copyright © 2012 Georgi Todorov  <terahz@geodar.com>
 //
 void PCA9685::set_pwm_freq(float freq) {
-  uint8_t prescale_val = (CLOCK_FREQ / 4096 / freq)  - 1;
+  uint8_t prescale_val = static_cast<uint8_t>(CLOCK_FREQ / 4096 / freq)  - 1;
   this->write_reg(Register::MODE1, 0x10); //sleep
   this->write_reg(Register::PRE_SCALE, prescale_val); // multiplyer for PWM frequency
   this->write_reg(Register::MODE1, 0x80); //restart
